% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangePoint.R
\name{ChangePoint}
\alias{ChangePoint}
\title{The Detection of Change Point using a MCMC method}
\usage{
ChangePoint(y, v1, v2, e1, e2, N, burnin)
}
\arguments{
\item{y}{a vector that contains the interested value in time series data.}

\item{v1}{the mean of the prior distribution of a Normal distribution(mu1,sigma^2)'s mu1.i.e. mu1 ~N(v1,e1^2).}

\item{v2}{the mean of the prior distribution of a Normal distribution(mu2,sigma^2)'s mu2.i.e. mu2 ~N(v2,e2^2).}

\item{e1}{the standard error of the prior distribution of a Normal distribution(mu1,sigma^2)'s mu1.i.e. mu1 ~N(v1,e1^2).}

\item{e2}{the standard error of the prior distribution of a Normal distribution(mu2,sigma^2)'s mu2.i.e. mu2 ~N(v2,e2^2).}

\item{N}{the length of chain.}

\item{burnin}{the burn-in time of the chain.}
}
\value{
a generated sample of possible change point
}
\description{
returns the generated vector of possible change point,to see the most likely one, please use a histogram or a kernel density estimation.
}
\examples{
\dontrun{
result<-ChangePoint(data)
hist(result,freq=FALSE,main="histogram of the change point",xlab="")
}
}
