% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Checkfunction2.R
\name{Checkfunction2}
\alias{Checkfunction2}
\title{The Detection of Change Point using an intuitive method}
\usage{
Checkfunction2(y, k, p)
}
\arguments{
\item{y}{a vector that contains the interested value in time series data.}

\item{k}{the k nearest point involved with the computation of slope.}

\item{p}{the mean of the prior distribution of a Normal distribution(mu2,sigma^2)'s mu2.i.e. mu2 ~N(v2,e2^2).}
}
\value{
possible change point
}
\description{
returns the the possible change point.
}
\examples{
\dontrun{
Checkfunction2(data,7,6)
}
}
